First pass at creating the files necessary to run python setup.py install
This will install pipeline_executor.py as a script and pipeline as a python package.
Install currently happens in whatever the default /usr path, unless the --home option is specified
Updates will be made in the future to improve the README, installation instructions (via INSTALL) and may add more sophisticated installation options.