First check-in (of several) to more generally handle location of pickled backup files.
Adds the ability to generally get and set location of backups based on output directory.
Uses current working directory if no output directory is specified.

NOTE: This code is broken, as location of backups is still hard-coded for pickling process. This will be fixed over the weekend or on Monday.

References issue #17.