First check-in of several expected to address issues #2 and #10.
Beginning of framework for client/server communications.
Added ability for pipeline_executor to function as both a client and a server.
Added clients array in pipeline so it can keep a list of currently registered clients.

IMPORTANT NOTE: In this current check-in, multiple executors will not work. This is a bug that will be fixed on the next check-in.