More changes to NLIN classes. 1. Creation of NLINBase to clean up duplicate code in subclasses. 2. Added comments/documentation throughout. 3. Added a final group to inputFH at the end of the model building, to be used in the stats class.
Added ability (in both NLIN.py and associated atoms) to override naming conventions.
i.e. We can specify an output to the minctracc, mincANTS or mincresample atoms that will override the default naming conventions while still using file handlers.
This partially addresses issues #45 and is somewhat related to issue #52.

Minor change: updated mincresample and autocrop to use output instead of outFile as an argument, for consistency with other classes.