Added command line options to LSQ6 option group so inormalize and the non-uniformity correction are now optional. (Includes related changes to LSQ6.py and MBM.py)
Added --no-calc-stats command line option to stats module, so --calc-stats does not need to be specified, it is now the default.

Fixed some small bugs with how masks are handled in mincresampleFileAndMask and NonUniformityCorrection. Masks are now handled appropriately if NUC and inormalize are not run but --mask-dir is specified at run time.