Updates to CalcStats and related calls so that kernels can be read in directly as a cmd line option.
This means that the repetitive code of putting blurs into an array in MBM.py, NLIN.py and registration_chain.py is now gone.

The setupBlurs function (introduced in CalcStats and LongitudinalStatsConcatAndResample) allows blurs to be read in
as either a comma separated string (as in the command line option) or as an array of doubles.

Part of Issue #82.