Added a hack (I HATE THIS!) into MBM.py so that prior to NLIN, if the initial target (created via LSQ12) has no mask, the init model mask is associated with this target. This is the default in old MBM.
I did this because otherwise when mincANTS is run, there is no target mask included in the registration. (Similarly for minctracc, but since we have a source mask there, it is less of an issue.)

Creating a github issue for this, so we can write this more seamlessly (e.g. by including it in either the NLIN or LSQ12 modules.)