First check-in of several, I'm sure. Main goal of this commit: add LSQ12 to second level of two level registration code.
This involved replacing initializeAndRunNLIN in the second level loop of twolevel_model_building.py with FullIterativeLSQ12Nlin. All fine and good, but this resulted in an error (since pyminc cannnot find the resolution of a file that does not yet exist) so I had to change a fair amount of underlying code. There is a discussion of this in Issue #76. Per discussion today with Jason, we should only try to get the file resolution using pyminc if we need it, which is often not the case.

1. Created returnFinestResolution function. This removes a lot of duplicate code, and returns a more sensible error when getting the file resolution fails. All error handling happens in this function.
2. In several functions, returnFinestResolution is only called if a registration protocol does not exist, since file resolution is usually used to set default protocols. In cases like LSQ6, it is always called, since right now, it is needed throughout the class.
3. Error handling in minc_parameters classes so a message is given if we try to set default protocols and file resolution does not exist.
4. minc_parameters now only set defaults if a protocol is NOT specified, so protocols now must specify all parameters instead of a subset.
5. small changes to FullIterativeLSQ12Nlin: changed order of arguments since initModel is optional and set lsq12_likeFile more sensibly.

Next step, which will set the stage for resolving Issue #94 and #76:
1. Ability to pass protocols from one module to another. E.g. We get a protocol from the command line and if a second protocol is not specified, we can pass the params (already parsed) to another module.
2. Add LSQ6 and LSQ12 into the first level of the two level code.