Fixed directory structure for twolevel code. Now, second_level_processed contains a sub-directory for each level one non-linear average, and all subsequent output files are placed here.
This necessitated an additional fix: adding a new parameter, avgPrefix, to all classes that call iterative non-linear registrations. This means that the default naming is no longer nlin-1.mnc, nlin-2.mnc, etc for non-linear averages, but prefix-nlin-1.mnc, prefix-nlin-2.mnc, etc. None may also be passed. For many registrations, the prefix is the pipeline name. For the first level of two level model building, it is subj-name-nlin-1.mnc, etc.

Also fixed an error with resampling first level determinants into common space. Was previously asking for the wrong (and non-existent!) transform.