Ben and I settled on a way to deal with FileAtoms and MincAtoms. In terms of creating similar output directory structures as the old code, for instance:
pipename_processed/file_1/tmp
pipename_processed/file_1/resampled
pipename_processed/file_1/stats-volumes

etc., a FileAtom/MincAtom now accepts a argument called pipeline_sub_dir. In the example above that would be the (full path) to pipename_processed. You can also specify a second argument called output_sub_dir, which in this case would be file_1. However, if this is not provided, the filename_wo_ext will be used for this.

When creating new objects from a FileAtom/MincAtom, a subdir argument can be specified which will place the new file in the appropriate directory (e.g., tmp, resampled).

Fixed the doctests in the code and removed obsolete comments.