use Jinja2 to generate most commands
We allow modifing commands via adding additional template directories
(in addition to pydpiper/templates) to PYDPIPER_TEMPLATE_PATH.
Currently we don't support template inheritance/overriding so
one must copy the entire original template.
In future we could also use templating for containerization and
general command wrappers.

Note that we currently do not support escaping (similar to Airflow's
BashOperator), but we should probably apply shlex.quote to interpolated
fragments automatically to handle files containing special characters
(from a usability, not security perspective - shlex.quote might
not be sufficiently secure and the user controls the templates anyway).
Also, we use shlex.split to generate an array to be passed to subprocess
but then execute the subprocess with shell=True on the executor.
Also, currently multiple bash commands are not supported in CmdStages
(an existing Pydpiper limitations) so we ignore newlines for ease of
writing.  We probably want to allow larger Bash chunks at some point
(similar to Nixpkgs builder scripts).