remove functionality to submit server to PBS queues
This was specific to PBS queues and not maintained.
It did provide useful functionality for queues where the server needs to
be run from the compute nodes but where the compute nodes cannot
themselves submit to the queue.
However, the new Makeflow backend is probably flexible enough to replace
this use case (e.g. by running the Makeflow process somewhere else and
submitting via SSH).